update materi controller

+++ src/main/java/backend/MateriController.java
@@ -10,6 +10,9 @@ import io.micronaut.http.annotation.Post;
 import io.micronaut.http.annotation.Delete;
 import io.micronaut.validation.Validated;
 import io.micronaut.http.MediaType;
+import io.micronaut.security.annotation.Secured;
+import io.micronaut.security.authentication.Authentication;
+import javax.annotation.Nullable;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonParser;
@@ -18,9 +21,11 @@ import com.google.gson.JsonObject;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.ArrayList;
 
 @Validated
 @Controller("/materi")
+@Secured("isAnonymous()")
 public class MateriController {
 
     private MateriRepository materiRepository;
@@ -45,15 +50,32 @@ public class MateriController {
     }
 
     @Post("/")
-    public String create(@Body Materi materi) {
+    @Secured("isAnonymous()")
+    public String create(@Body Materi materi, @Nullable Authentication authentication) {
+
 
         try {
 
-            Materi result = materiRepository.save(materi);
-                
-            MateriResponse response = new MateriResponse("ok", "Berhasil menambahkan data materi", result);
+            if(authentication == null) {
+                MateriResponse response = new MateriResponse("error", "Unauthorized user.");
 
-            return new Gson().toJson(response);
+                return new Gson().toJson(response);
+            } else {
+                Object data = authentication.getAttributes().get("roles");
+                String roles = data.toString();
+
+                if(roles.equals("[\"Admin\"]")) {
+                    Materi result = materiRepository.save(materi);
+                    
+                    MateriResponse response = new MateriResponse("ok", "Berhasil menambahkan data materi", result);
+
+                    return new Gson().toJson(response);
+                } else {
+                    MateriResponse response = new MateriResponse("error", "Anda tidak boleh mengakses halaman ini.");
+
+                    return new Gson().toJson(response);
+                }
+
             
 
         } catch(Exception e) {
@@ -66,6 +88,7 @@ public class MateriController {
     }
 
     @Get("/{id}")
+    @Secured("isAnonymous()")
     public String show(Long id) {
 
         try {
@@ -94,37 +117,70 @@ public class MateriController {
     }
 
     @Post("/{id}") 
-    public String update(@Body Materi materi) {
-        // Materi getMateri = materiRepository.findById(id);
-
-        Materi result = materiRepository.update(materi);
-
-        if(result != null) {
-            MateriResponse response = new MateriResponse("ok", "Berhasil memperbarui data materi", result);
+    @Secured("isAnonymous()")
+    public String update(@Body Materi materi, @Nullable Authentication authentication) 
+    {
+        
+        if(authentication == null) {
+            MateriResponse response = new MateriResponse("error", "Unauthorized user.");
 
             return new Gson().toJson(response);
         } else {
-            MateriResponse response = new MateriResponse("error", "Data materi tidak ditemukan");
+            Object data = authentication.getAttributes().get("roles");
+            String roles = data.toString();
 
-            return new Gson().toJson(response);
+            if(roles.equals("[\"Admin\"]")) {
+                Materi result = materiRepository.update(materi);
+
+                if(result != null) {
+                    MateriResponse response = new MateriResponse("ok", "Berhasil memperbarui data materi", result);
+                    Materi result = materiRepository.update(materi);
+
+                    return new Gson().toJson(response);
+                } else {
+                    MateriResponse response = new MateriResponse("error", "Data materi tidak ditemukan");
+
+                    return new Gson().toJson(response);
+                }
+            } else {
+                MateriResponse response = new MateriResponse("error", "Anda tidak boleh mengakses halaman ini.");
+                return new Gson().toJson(response);
+            } 
         }
     }
 
+
     @Delete("/{id}")
-    public String delete(Long id) {
-        Materi getMateri = materiRepository.findById(id);
-         
-        if(getMateri != null) {
-            materiRepository.deleteById(id);
+    @Secured("isAnonymous()")
+    public String delete(Long id, @Nullable Authentication authentication) 
+    {
 
-            MateriResponse response = new MateriResponse("ok", "Berhasil menghapus data materi");
-
+        if(authentication == null) {
+            MateriResponse response = new MateriResponse("error", "Unauthorized user.");
             return new Gson().toJson(response);
-
         } else {
-            MateriResponse response = new MateriResponse("error", "Data materi tidak ditemukan");
+            Object data = authentication.getAttributes().get("roles");
+            String roles = data.toString();
 
-            return new Gson().toJson(response);
+            if(roles.equals("[\"Admin\"]")) {
+                Materi getMateri = materiRepository.findById(id);
+                         
+                if(getMateri != null) {
+                    materiRepository.deleteById(id);
+
+                    MateriResponse response = new MateriResponse("ok", "Berhasil menghapus data materi");
+
+                    return new Gson().toJson(response);
+
+                } else {
+                    MateriResponse response = new MateriResponse("error", "Data materi tidak ditemukan");
+
+                    return new Gson().toJson(response);
+                }
+            } else {
+                MateriResponse response = new MateriResponse("error", "Anda tidak boleh mengakses halaman ini.");
+                return new Gson().toJson(response);
+            }
         }
     }
 }
